Extract shared `assertCompilationError`/`muteCompilationError` test fixtures
Introduce two reusable test fixtures in `:base`:
 - `muteCompilationError` runs a block with console output muted (delegating
   to `tapConsole`), so deliberately provoked compilation errors do not pollute
   the build log;
 - `assertCompilationError` composes it with `assertThrows<Compilation.Error>`
   and returns the thrown error.

Replace the `assertThrows<Compilation.Error> { tapConsole { ... } }` pattern with
`assertCompilationError { ... }` across the `base`, `entity-tests`, `signal-tests`,
and `comparable-tests` modules for more compact test code.

Move `Logging.testLib` to `testFixturesImplementation` in `:base` (the fixtures
need `tapConsole`); consumer modules now obtain it transitively via
`testFixtures(:base)`, so drop their redundant declarations.
