+++ comparable-tests/build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025, TeamDev. All rights reserved.
+ * Copyright 2026, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,7 +25,6 @@
  */
 
 import io.spine.dependency.lib.Protobuf
-import io.spine.dependency.local.Logging
 import io.spine.dependency.local.TestLib
 
 plugins {
@@ -36,7 +35,6 @@ plugins {
 
 dependencies {
     testImplementation(TestLib.lib)
-    testImplementation(Logging.testLib)?.because("We need `tapConsole`.")
     testImplementation(project(":comparable"))
     testImplementation(testFixtures(project(":base")))
 }
