+++ entity-tests/build.gradle.kts
@@ -25,7 +25,6 @@
  */
 
 import io.spine.dependency.lib.Protobuf
-import io.spine.dependency.local.Logging
 
 plugins {
     module
@@ -41,7 +40,6 @@ dependencies {
     ).forEach {
         testImplementation(it)
     }
-    testImplementation(Logging.testLib)?.because("We need `tapConsole`.")
 }
 
 protobuf {
