+++ signal-tests/src/test/kotlin/io/spine/tools/core/jvm/signal/UnsupportedCommandIdTypeErrorSpec.kt
@@ -27,14 +27,12 @@
 package io.spine.tools.core.jvm.signal
 
 import io.kotest.matchers.string.shouldContain
-import io.spine.logging.testing.tapConsole
 import io.spine.testing.compiler.acceptingOnly
-import io.spine.tools.compiler.Compilation
+import io.spine.tools.core.jvm.assertCompilationError
 import io.spine.tools.core.signal.given.command.RepeatedIdCommand
 import java.nio.file.Path
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.api.io.TempDir
 
 /**
@@ -56,11 +54,8 @@ internal class UnsupportedCommandIdTypeErrorSpec {
     @Test
     fun `reject the target-entity ID field of a 'repeated' type`(@TempDir projectDir: Path) {
         val descriptor = RepeatedIdCommand.getDescriptor()
-        val error = assertThrows<Compilation.Error> {
-            // Mute the expected compilation error so it does not surface in the build log.
-            tapConsole {
-                runPipeline(projectDir, acceptingOnly(descriptor))
-            }
+        val error = assertCompilationError {
+            runPipeline(projectDir, acceptingOnly(descriptor))
         }
         error.message.let {
             it shouldContain "${descriptor.fullName}.telescope"
